// src/shared/tokens.generated.ts
export const TOKENS: TokenTable = {
	"themes": [
		"custom-light",
		"empeo-dark",
		"empeo-light",
		"etaxgo-light",
		"salesbear-light",
		"venio-dark",
		"venio-light"
	],
	"rows": [
		{
			"key": "background-default",
			"cssVar": "--go5-background-default",
			"cls": null,
			"group": "other",
			"values": {
				"custom-light": "#F5F5F5",
				"empeo-dark": "#383842",
				"empeo-light": "#ECECF1",
				"etaxgo-light": "#ECECF1",
				"salesbear-light": "#F5F5F5",
				"venio-dark": "#383842",
				"venio-light": "#ECECF1"
			}
		},
// ...
};